spilt cha_bloc into three files
