+++ lib/src/features/chat/presentation/bloc/archive_event.dart
@@ -1,0 +1,15 @@
+abstract class ChatEvent {}
+
+class LoadChatsEvent extends ChatEvent {}
+
+class LoadArchivedChatsEvent extends ChatEvent {}
+
+class ArchiveChatEvent extends ChatEvent {
+  final String chatId;
+  ArchiveChatEvent(this.chatId);
+}
+
+class DeleteChatEvent extends ChatEvent {
+  final String chatId;
+  DeleteChatEvent(this.chatId);
+}
